refactor : OutputView 객체화 및 출력 기능 분화 - static 메서드 non-static화 - OutputView의 print 역할 캡슐화 - content 전달시 개행 방식을 인자로 함께 전달하여 다형성 확보

+++ src/main/java/bridge/domain/result/ResultRendering.java
@@ -43,7 +43,7 @@ public class ResultRendering {
 		return RESULT_PRESENTATION + ENTER
 				+ BridgeDescription + ENTER
 				+ String.format(RESULT_FAIL_OR_SUCCESS, gameResult.keySet().toArray()[0]) + ENTER
-				+ String.format(RESULT_TRIAL, gameResult.values().toArray()[0]);
+				+ String.format(RESULT_TRIAL, (Integer) gameResult.values().toArray()[0]);
 	}
 
 	private void buildDescription(String crossingMark, List<String> passingSide, List<String> nonPassingSide) {
